Airway-18

This function processes and analyzes the dataset to calculate the "Airway-18" NEMSQA metric. It includes cleaning and transforming several columns related to patient data, airway procedures, and vital signs, and it returns a cleaned dataset with the relevant calculations. The final calculation is an assessment of the successful last invasive airway procedures performed during an EMS response originating from a 911 request in which waveform capnography is used for tube placement confirmation.

airway_18(
  df = NULL,
  patient_scene_table = NULL,
  procedures_table = NULL,
  vitals_table = NULL,
  airway_table = NULL,
  response_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  eprocedures_01_col,
  eprocedures_02_col,
  eprocedures_03_col,
  eprocedures_05_col,
  eprocedures_06_col,
  eairway_02_col = NULL,
  eairway_04_col = NULL,
  evitals_01_col,
  evitals_16_col,
  ...
)

Arguments

df

A data frame or tibble containing the dataset to be processed. Default is NULL.

patient_scene_table

A data frame or tibble containing only ePatient and eScene fields as a fact table. Default is NULL.

procedures_table

A data frame or tibble containing only the eProcedures fields needed for this measure's calculations. Default is NULL.

vitals_table

A data frame or tibble containing only the eVitals fields needed for this measure's calculations. Default is NULL.

airway_table

A data frame or tibble containing only the eAirway fields needed for this measure's calculations. Default is NULL.

response_table

A data frame or tibble containing only the eResponse fields needed for this measure's calculations. Default is NULL.

erecord_01_col

<'tidy-select'> Column name containing the unique patient record identifier.

incident_date_col

<'tidy-select'> Column name containing the incident date. Default is NULL.

epatient_15_col

<'tidy-select'> Column name for patient information (exact purpose unclear).

epatient_16_col

<'tidy-select'> Column name for patient information (exact purpose unclear).

eresponse_05_col

<'tidy-select'> Column name for emergency response codes.

eprocedures_01_col

<'tidy-select'> Column name for procedure times or other related data.

eprocedures_02_col

<'tidy-select'> Column name for additional procedure data.

eprocedures_03_col

<'tidy-select'> Column name for procedure codes.

eprocedures_05_col

<'tidy-select'> Column name for number of procedure attempts.

eprocedures_06_col

<'tidy-select'> Column name for procedure success codes.

eairway_02_col

<'tidy-select'> Column name for airway procedure data (datetime). Default is NULL.

eairway_04_col

<'tidy-select'> Column name for airway procedure data. Default is NULL.

evitals_01_col

<'tidy-select'> Column name for vital signs data (datetime).

evitals_16_col

<'tidy-select'> Column name for additional vital signs data.

...

Additional arguments passed to other functions if needed.

patient_dob_col

<'tidy-select'> Column name containing the patient's date of birth. Default is NULL.

Value

A tibble summarizing results for Adults and Peds with the following columns: pop: Population type (Adults, Peds). numerator: Count of incidents where waveform capnography is used for tube placement confirmation on the last successful invasive airway procedure. denominator: Total count of incidents. prop: Proportion of incidents where waveform capnography is used for tube placement confirmation on the last successful invasive airway procedure. prop_label: Proportion formatted as a percentage with a specified number of decimal places.

Data Assumptions

This function assumes that: Data are already loaded. The data needs to be a data.frame or tibble.

Age in years can be calculated using the patient date of birth and incident date. These fields must have valid Date or POSIXct data types. If patient DOB and the incident date are not passed, the function will use the system generated patient age and age units.

When values are missing, they are coded as NA, not the "not known"/"not recorded" values common to ImageTrend or the NEMSIS codes that correspond to "not values".

The date time fields from eProcedures.01, eAirway.02, and eVitals.01 must be date time objects or the function will fail.

The eAirway fields are optional in this function in case a user does not have access to those fields.

Practical Tips

Ensure data are pre-processed, with missing values coded as NA, before passing into the function. Prepare necessary joins (e.g., for vitals) in advance; this function does not perform joins.

Author

Nicolas Foss, Ed.D., MS