Asthma-01 Populations

Filters data down to the target populations for Asthma-01, and categorizes records to identify needed information for the calculations.

Identifies key categories related to asthma-related incidents in an EMS dataset, specifically focusing on cases where 911 was called for respiratory distress, and certain medications were administered. This function segments the data by age into adult and pediatric populations, computing the proportion of cases that received beta-agonist treatment.

asthma_01_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  medications_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  emedications_03_col
)

Arguments

df

A data.frame or tibble containing EMS data. Default is NULL.

patient_scene_table

A data.frame or tibble containing at least ePatient and eScene fields as a fact table. Default is NULL.

response_table

A data.frame or tibble containing at least the eResponse fields needed for this measure's calculations. Default is NULL.

situation_table

A data.frame or tibble containing at least the eSituation fields needed for this measure's calculations. Default is NULL.

medications_table

A data.frame or tibble containing at least the eMedications fields needed for this measure's calculations. Default is NULL.

erecord_01_col

<'tidy-select'> The column representing the EMS record unique identifier. Default is NULL.

incident_date_col

<'tidy-select'> Column that contains the incident date. Default is NULL.

patient_DOB_col

<'tidy-select'> Column that contains the patient's date of birth. Default is NULL.

epatient_15_col

<'tidy-select'> Column representing the patient's numeric age agnostic of unit.

epatient_16_col

<'tidy-select'> Column representing the patient's age unit ("Years", "Months", "Days", "Hours", or "Minute").

eresponse_05_col

<'tidy-select'> Column that contains eResponse.05.

esituation_11_col

<'tidy-select'> Column that contains eSituation.11.

esituation_12_col

<'tidy-select'> Column that contains all eSituation.12 values as a single comma-separated list.

emedications_03_col

<'tidy-select'> Column that contains all eMedications.03 values as a single comma-separated list.

Value

#' A list that contains the following:

a tibble with counts for each filtering step,

a tibble for each population of interest

a tibble for the initial population

Data Assumptions

This function assumes that: Data are already loaded. The data need to be a data.frame or tibble.

Age in years can be calculated using the patient date of birth and incident date. These fields must have valid Date or POSIXct data types. If patient DOB and the incident date are not passed, the function will use the system generated patient age and age units.

When values are missing, they are coded as NA, not the "not known"/"not recorded" values common to ImageTrend or the NEMSIS codes that correspond to "not values".

The primary and secondary impression fields (eSituation.11 and eSituation.12) can have the ICD-10 codes and/or the text descriptions present in them. Similarly, this function assumes that the eResponse.05 column has NEMSIS codes and/or text descriptions present, but text can also be included for reference.

The eMedications.03 field contains all medications administered and contains a text description of the medication using the generic name. The Rx code may also be included for reference, and will be checked.

The secondary impressions field (eSituation.12) is best prepared as a comma-separated list of all values in a single string.

Practical Tips

The first argument is the data.frame or tables with data elements prepared as above. No joining is done. Any joins to get vitals, etc. will need to be done outside of this function.

Author

Nicolas Foss, Ed.D., MS