nemsqar: Streamlines NEMSQA Metric Calculations

nemsqar is an R package designed to automate the calculation of EMS quality metrics as defined by the National EMS Quality Alliance (NEMSQA). With a focus on enhancing EMS service performance through reliable, evidence-based quality measures, nemsqar aims to streamline metric calculation for healthcare providers, EMS agencies, and researchers dedicated to improving patient outcomes.

See also

Author

Maintainer: Nicolas Foss nicolas.foss@hhs.iowa.gov [copyright holder]

Authors:

Alyssa Green awhim@nemsqa.org

Samuel Kordik samuel.kordik@utsouthwestern.edu (ORCID)