Synthetic Test Data for Airway-18 Measure in National EMS Information System

This dataset provides completely synthetic test data for evaluating the airway_01 and airway_01_population functions in the nemsqar package. It contains fields relevant to airway management and procedural success, allowing users to test the functionality of these measure functions.

Usage

data(nemsqar_airway_01_df)

Format

A data frame with 1,000 rows and 16 variables:

Incident Patient Care Report Number - PCR (eRecord.01)

Unique identifier for the incident report (character).

Incident Date

Date of the incident (Date).

Patient Age (ePatient.15)

Age of the patient (numeric).

Patient Age Units (ePatient.16)

Unit of measurement for patient age (character).

Patient Date Of Birth (ePatient.17)

Patient's date of birth (Date).

Patient Gender (ePatient.13)

Gender of the patient (character).

Response Type Of Service Requested With Code (eResponse.05)

Type of service requested for the EMS response (character).

Cardiac Arrest During EMS Event With Code (eArrest.01)

Indicates whether a cardiac arrest occurred during the EMS event (character).

Vitals Signs Taken Date Time (eVitals.01)

Timestamp when vital signs were recorded (datetime).

Vitals Systolic Blood Pressure SBP (eVitals.06)

Recorded systolic blood pressure (numeric).

Vitals Pulse Oximetry (eVitals.12)

Oxygen saturation level (numeric).

Procedure Performed Date Time (eProcedures.01)

Timestamp when the procedure was performed (datetime).

Procedure Performed Prior To EMS Care (eProcedures.02)

Indicates if the procedure was performed before EMS arrival (character).

Procedure Performed Description And Code (eProcedures.03)

Description and standardized code of the performed procedure (character).

Procedure Number Of Attempts (eProcedures.05)

Number of attempts made for the procedure (numeric).

Procedure Successful (eProcedures.06)

Indicates whether the procedure was successful (character).

Details

Users are encouraged to test these functions with this dataset, but results should not be interpreted as meaningful. Some outputs may be nonsensical, which is expected since this data is only intended to demonstrate the expected structure of input data.

Examples

data(nemsqar_airway_01_df)
head(nemsqar_airway_01_df)
#> # A tibble: 6 × 16
#>   Incident Patient Care Report Number -…¹ `Incident Date` Patient Age (ePatien…²
#>   <chr>                                   <date>                           <dbl>
#> 1 NyXFBlJfnm-8333586176                   2023-02-22                           9
#> 2 XTLCINMLTP-8616021114                   2023-09-24                          66
#> 3 HfYjlIEQSk-9529756610                   2023-05-13                          27
#> 4 MOwVDhriyC-5915613206                   2023-09-05                         102
#> 5 ZCGOtLEPKw-7820135532                   2023-08-24                          36
#> 6 fEMvUCQCRQ-9052388486                   2023-02-25                          33
#> # ℹ abbreviated names:
#> #   ¹​`Incident Patient Care Report Number - PCR (eRecord.01)`,
#> #   ²​`Patient Age (ePatient.15)`
#> # ℹ 13 more variables: `Patient Age Units (ePatient.16)` <chr>,
#> #   `Patient Date Of Birth (ePatient.17)` <date>,
#> #   `Patient Gender (ePatient.13)` <chr>,
#> #   `Response Type Of Service Requested With Code (eResponse.05)` <chr>, …