Pediatrics-03B Populations

Filters data down to the target populations for Pediatrics-03B, and categorizes records to identify needed information for the calculations.

Identifies key categories related to diabetes/hypoglycemia incidents in an EMS dataset, specifically focusing on cases where 911 was called for diabetes/hypoglycemia distress, certain medications were administered, and a weight is taken. This function segments the data into pediatric populations, computing the proportion of cases that have a documented weight.

Usage

pediatrics_03b_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  exam_table = NULL,
  medications_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_dob_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  eexam_01_col,
  eexam_02_col,
  emedications_03_col,
  emedications_04_col
)

Arguments

df

A data frame or tibble containing emergency response records. Default is NULL.

patient_scene_table

A data.frame or tibble containing only ePatient and eScene fields as a fact table. Default is NULL.

response_table

A data.frame or tibble containing only the eResponse fields needed for this measure's calculations. Default is NULL.

exam_table

A data.frame or tibble containing only the eExam fields needed for this measure's calculations. Default is NULL.

medications_table

A data.frame or tibble containing only the eMedications fields needed for this measure's calculations. Default is NULL.

erecord_01_col

Column for unique EMS record identifiers.

incident_date_col

Column that contains the incident date. This defaults to NULL as it is optional in case not available due to PII restrictions.

patient_dob_col

Column that contains the patient's date of birth. This defaults to NULL as it is optional in case not available due to PII restrictions.

epatient_15_col

Column giving the calculated age value.

epatient_16_col

Column giving the provided age unit value.

eresponse_05_col

Column containing the EMS response codes.

eexam_01_col

Column containing documented weight information.

eexam_02_col

Another column for weight documentation, if applicable.

emedications_03_col

Column indicating medication administration.

emedications_04_col

Column listing medications administered.

Value

#' A list that contains the following:

a tibble with counts for each filtering step,

a tibble for each population of interest

a tibble for the initial population

Author

Nicolas Foss, Ed.D., MS