Respiratory-02 Populations

The respiratory_02_population function calculates metrics for pediatric and adult respiratory populations based on pre-defined criteria, such as low oxygen saturation and specific medication or procedure codes. It returns a summary table of the overall, pediatric, and adult populations, showing counts and proportions.

respiratory_02_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  vitals_table = NULL,
  medications_table = NULL,
  procedures_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  evitals_12_col,
  emedications_03_col,
  eprocedures_03_col
)

Arguments

df

A data frame containing incident data with each row representing an observation.

patient_scene_table

A data.frame or tibble containing at least epatient and escene fields as a fact table.

response_table

A data.frame or tibble containing at least the eresponse fields needed for this measure's calculations.

vitals_table

A data.frame or tibble containing at least the evitals fields needed for this measure's calculations.

medications_table

A data.frame or tibble containing only the emedications fields needed for this measure's calculations.

procedures_table

A data.frame or tibble containing only the eprocedures fields needed for this measure's calculations.

erecord_01_col

<'tidy-select'> Column name for eRecord.01, used to form a unique patient ID.

incident_date_col

<'tidy-select'> Date or POSIXct Column name for the incident date.

patient_DOB_col

<'tidy-select'> Date or POSIXct Column name for the patient's date of birth.

epatient_15_col

<'tidy-select'> integer Column giving the calculated age value.

epatient_16_col

<'tidy-select'> Column giving the provided age unit value.

eresponse_05_col

<'tidy-select'> Column name for response codes (e.g., incident type).

evitals_12_col

<'tidy-select'> Column name for oxygen saturation (SpO2) values.

emedications_03_col

<'tidy-select'> Column name for medication codes.

eprocedures_03_col

<'tidy-select'> Column name for procedure codes.

Value

#' A list that contains the following:

a tibble with counts for each filtering step,

a tibble for each population of interest

a tibble for the initial population

Data Assumptions

Assume data are already loaded The data must be a dataframe or tibble that contains emedications.03 and eprocedures.03 as columns where each cell contains all values entered for each respective incident. These are not list columns but can be comma separated values in each cell, and must contain all medications for each incident. The table must also have the erecord.01 values to attempt a best estimate of a unique ID by concatenating the erecord.01, incident date, and patient DOB. The table can include all values entered for evitals.12 in the table. This will cause row explosion, but the function will find if the patient had sp02 < 90 among any values entered. This function assumes you have the following fields: eRecord.01, eresponse.05 (must contain code), evitals.12 (all values), emedications.03 (all with code), eprocedures.03 (with code) this function will calculate age in years for the end user. this function also assumes that rows that are missing any value are NA, not the not known / not recorded values common to ImageTrend or the value codes that correspond to "not values". the first argument is a dataframe, no joining is done in the function. any needed table joins must be done prior to running the function to ensure required columns are present. Grouping by specific attributes (e.g., region) can be performed inside this function by utilizing the .by argument passed via tidydots (i.e. ...) to dplyr::summarize.

Author

Nicolas Foss, Ed.D., MS