Seizure-02 Calculation

Calculates the NEMSQA Seizure-02 Measure.

seizure_02(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  medications_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  emedications_03_col,
  ...
)

Arguments

df

A data frame where each row is an observation, containing all necessary columns for analysis.

patient_scene_table

A data frame or tibble containing only epatient and escene fields as a fact table. Default is NULL.

response_table

A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is NULL.

situation_table

A data.frame or tibble containing only the esituation fields needed for this measure's calculations. Default is NULL.

medications_table

A data.frame or tibble containing only the emedications fields needed for this measure's calculations. Default is NULL.

erecord_01_col

<'tidy-select'> The column containing unique record identifiers for each encounter.

incident_date_col

<'tidy-select'> Column name representing the incident date as Date or POSIXct.

patient_DOB_col

<'tidy-select'> Column name representing the patient's date of birth as Date or POSIXct.

epatient_15_col

<'tidy-select'> Column name for patient age in numeric form.

epatient_16_col

<'tidy-select'> Column name for age unit (e.g., "Years" or "Months").

eresponse_05_col

<'tidy-select'> Column name for response codes; "911" call codes are filtered.

esituation_11_col

<'tidy-select'> Column name for primary impressions.

esituation_12_col

<'tidy-select'> Column name for secondary impressions.

emedications_03_col

<'tidy-select'> Column name for medications administered; ideally a list column or string with comma-separated values.

...

Additional arguments passed to dplyr::summarize.

Value

A tibble summarizing results for three population groups (All, Adults, and Peds) with the following columns: pop: Population type (All, Adults, or Peds). numerator: Count of incidents where beta-agonist medications were administered. denominator: Total count of incidents. prop: Proportion of incidents involving beta-agonist medications. prop_label: Proportion formatted as a percentage with a specified number of decimal places.

Details

Calculates age-based seizure metrics for a dataset. This function filters data for patients based on incident information, diagnoses, and administered medications to assess adherence to Seizure-02 metrics.

Data Assumptions

This function assumes that:

Data are already loaded in a data frame or tibble where each row represents one observation (e.g., patient), and each column is a feature (field). Alternatively, data can consist of distinct datasets referenced as unique columns.

Age in years will be calculated using the patient's date of birth and the incident date. These fields must be valid Date or POSIXct data types.

Any missing values in rows are coded as NA rather than ImageTrend-specific "not known" or "not recorded" values, or other codes that correspond to unspecified values.

The primary and secondary impression fields contain ICD-10 codes, although descriptive text may also be present for reference. These fields are identified as eSituation.11 for the primary impression and eSituation.12 for secondary impressions.

The eResponse.05 column contains NEMSIS response codes. Additional descriptive text may also be present in this field for reference.

The eMedications.03 column holds all medications administered to a patient during an event, represented as a list column or comma-separated values. Each entry should use generic medication names, and while RxNORM codes can be included, they will not be validated by the function.

The eSituation.12 (secondary impressions) field is best prepared as a comma-separated list within a single string or as a list column containing each secondary impression.

The first argument to the function is a data frame; any joins needed to incorporate additional data (e.g., vitals) must be conducted prior to calling this function.

Grouping by specific attributes (e.g., region) can be performed inside this function by utilizing the .by argument passed via tidydots (i.e. ...) to dplyr::summarize.

Practical Tips

The first argument is the data.frame prepared as above. No joining is done. Any joins to get specific data elements will need to be done outside of this function.

Value

A summarized tibble with counts and proportions of benzodiazepine treatment among 911 seizure calls, segmented by population group.

Author

Nicolas Foss, Ed.D., MS