Stroke-01 Calculation

The stroke_01 function processes EMS dataset to identify potential stroke cases based on specific criteria and calculates the stroke scale measures. It filters the data for 911 response calls, identifies stroke-related impressions and scales, and aggregates results by unique patient encounters.

Usage

stroke_01(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  evitals_23_col,
  evitals_26_col,
  evitals_29_col,
  evitals_30_col,
  ...
)

Arguments

df

A data frame or tibble containing the dataset. Each row should represent a unique patient encounter.

patient_scene_table

A data frame or tibble containing only epatient and escene fields as a fact table. Default is NULL.

response_table

A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is NULL.

situation_table

A data.frame or tibble containing only the esituation fields needed for this measure's calculations. Default is NULL.

vitals_table

A data.frame or tibble containing only the evitals fields needed for this measure's calculations. Default is NULL.

erecord_01_col

The column containing unique record identifiers for each encounter.

eresponse_05_col

The column containing EMS response codes, which should include 911 response codes.

esituation_11_col

The column containing the primary impression codes or descriptions related to the situation.

esituation_12_col

The column containing secondary impression codes or descriptions related to the situation.

evitals_23_col

The column containing the Glasgow Coma Scale (GCS) score.

evitals_26_col

The column containing the AVPU (alert, verbal, pain, unresponsive) scale value.

evitals_29_col

The column containing the stroke scale score achieved during assessment.

evitals_30_col

The column containing stroke scale type descriptors (e.g., FAST, NIH, etc.).

...

Additional arguments passed to dplyr::summarize() function for further customization of results.

Value

A tibble summarizing results for the total population with the following columns: pop: Population type (All). numerator: Count of incidents where beta-agonist medications were administered. denominator: Total count of incidents. prop: Proportion of incidents involving beta-agonist medications. prop_label: Proportion formatted as a percentage with a specified number of decimal places.

Author

Nicolas Foss, Ed.D., MS