Stroke-01 Populations

Filters data down to the target populations for Stroke-01, and categorizes records to identify needed information for the calculations.

stroke_01_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  evitals_23_col,
  evitals_26_col,
  evitals_29_col,
  evitals_30_col
)

Arguments

df

A data frame or tibble containing the dataset. Each row should represent a unique patient encounter.

patient_scene_table

A data frame or tibble containing only epatient and escene fields as a fact table. Default is NULL.

response_table

A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is NULL.

situation_table

A data.frame or tibble containing only the esituation fields needed for this measure's calculations. Default is NULL.

vitals_table

A data.frame or tibble containing only the evitals fields needed for this measure's calculations. Default is NULL.

erecord_01_col

<'tidy-select'> The column containing unique record identifiers for each encounter.

eresponse_05_col

<'tidy-select'> The column containing EMS response codes, which should include 911 response codes.

esituation_11_col

<'tidy-select'> The column containing the primary impression codes or descriptions related to the situation.

esituation_12_col

<'tidy-select'> The column containing secondary impression codes or descriptions related to the situation.

evitals_23_col

<'tidy-select'> The column containing the Glasgow Coma Scale (GCS) score.

evitals_26_col

<'tidy-select'> The column containing the AVPU (alert, verbal, pain, unresponsive) scale value.

evitals_29_col

<'tidy-select'> The column containing the stroke scale score achieved during assessment.

evitals_30_col

<'tidy-select'> The column containing stroke scale type descriptors (e.g., FAST, NIH, etc.).

Value

#' A list that contains the following:

a tibble with counts for each filtering step,

a tibble for each population of interest

a tibble for the initial population

Details

Identifies key categories related to stroke-related incidents in an EMS dataset, specifically focusing on cases where 911 was called for stroke, and a stroke scale was administered. .

Data Assumptions

This function assumes that:

Data are already loaded into a data frame or tibble where each row represents one observation (e.g., patient) and each column is a dplyr::distinct feature (field). Alternatively, data may consist of separate datasets referenced by unique columns.

Missing values in rows are represented as NA. "Not known" or "not recorded" values, common to ImageTrend or other non-numeric codes, should be pre-cleaned before use.

The eresponse_05_col contains NEMSIS response codes and may also include descriptive text for reference.

The vital signs columns evitals.23, evitals.26, evitals.29, and evitals.30 can contain all the unique values entered per incident, which may cause row explosion. The function will handle this issue elegantly under the hood.

The esituation.11 contains single responses per response, but esituation.12 should be a list column or text field with all relevant secondary provider impression entries provided in a single cell as comma-separated values for each unique incident.

The first argument to this function is the main data frame. No joins are performed within the function; any necessary joins (e.g., to incorporate vitals or additional fields) should be completed prior to calling this function.

Grouping by specific attributes (e.g., region) can be performed inside this function by utilizing the .by argument passed via tidydots (i.e. ...) to dplyr::summarize.

Practical Tips

Ensure data are pre-processed, with missing values coded as NA, before passing into the function. Prepare necessary joins (e.g., for vitals) in advance; this function does not perform joins.

Author

Nicolas Foss, Ed.D., MS