Trauma-01 Calculation

This function processes EMS data to calculate the Trauma-01 performance measure, which evaluates the percentage of trauma patients assessed for pain using a numeric scale. The function filters and summarizes the data based on specified inclusion criteria.

Usage

trauma_01(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  disposition_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  esituation_02_col,
  eresponse_05_col,
  evitals_23_col,
  evitals_26_col,
  evitals_27_col,
  edisposition_28_col,
  transport_disposition_col,
  ...
)

Arguments

df

A data frame or tibble containing EMS records. Default is NULL.

patient_scene_table

A data frame or tibble containing only epatient and escene fields as a fact table. Default is NULL.

response_table

A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is NULL.

situation_table

A data.frame or tibble containing only the esituation fields needed for this measure's calculations. Default is NULL.

disposition_table

A data.frame or tibble containing only the edisposition fields needed for this measure's calculations. Default is NULL.

vitals_table

A data.frame or tibble containing only the evitals fields needed for this measure's calculations. Default is NULL.

erecord_01_col

Column name representing the EMS record ID.

incident_date_col

Column that contains the incident date. This defaults to NULL as it is optional in case not available due to PII restrictions.

patient_DOB_col

Column that contains the patient's date of birth. This defaults to NULL as it is optional in case not available due to PII restrictions.

epatient_15_col

Column name for the patient's age in numeric format.

epatient_16_col

Column name for the unit of age (e.g., "Years", "Months").

esituation_02_col

Column name indicating if the situation involved an injury.

eresponse_05_col

Column name for the type of EMS response (e.g., 911 call).

evitals_23_col

Column name for the Glasgow Coma Scale (GCS) total score.

evitals_26_col

Column name for AVPU (Alert, Voice, Pain, Unresponsive) status.

evitals_27_col

Column name for the pain scale assessment.

edisposition_28_col

Column name for patient care disposition details.

transport_disposition_col

Column name for transport disposition details.

...

Additional arguments passed to the dplyr::summarize function for custom summarization.

Value

A tibble summarizing results for three population groups (All, Adults, and Peds) with the following columns:

measure: The name of the measure being calculated. pop: Population type (All, Adults, Peds). numerator: Count of incidents where a pain scale was administered. denominator: Total count of incidents. prop: Proportion of incidents where a pain scale was administered. prop_label: Proportion formatted as a percentage with a specified number of decimal places.

Author

Nicolas Foss, Ed.D., MS

Examples

# Synthetic test data
  test_data <- tibble::tibble(
    erecord_01 = c("R1", "R2", "R3", "R4", "R5"),
    epatient_15 = c(34, 5, 45, 2, 60),  # Ages
    epatient_16 = c("Years", "Years", "Years", "Months", "Years"),
    eresponse_05 = rep(2205001, 5),
    esituation_02 = rep("Yes", 5),
    evitals_23 = rep(15, 5),
    evitals_26 = rep("Alert", 5),
    evitals_27 = c(0, 2, 4, 6, 8),
    edisposition_28 = rep(4228001, 5),
    edisposition_30 = c(4230001, 4230003, 4230001, 4230007, 4230007)
  )

  # Run function
  trauma_01(
    df = test_data,
    erecord_01_col = erecord_01,
    epatient_15_col = epatient_15,
    epatient_16_col = epatient_16,
    eresponse_05_col = eresponse_05,
    esituation_02_col = esituation_02,
    evitals_23_col = evitals_23,
    evitals_26_col = evitals_26,
    evitals_27_col = evitals_27,
    edisposition_28_col = edisposition_28,
    transport_disposition_col = edisposition_30
  )
#> 
#> ── Trauma-01 ───────────────────────────────────────────────────────────────────
#> 
#> ── Gathering Records for Trauma-01 ──
#> 
#> Running `trauma_01_population()`  [Working on 1 of 14 tasks] ●●●───────────────…
#> Running `trauma_01_population()`  [Working on 2 of 14 tasks] ●●●●●─────────────…
#> Running `trauma_01_population()`  [Working on 3 of 14 tasks] ●●●●●●●───────────…
#> Running `trauma_01_population()`  [Working on 4 of 14 tasks] ●●●●●●●●●●────────…
#> Running `trauma_01_population()`  [Working on 5 of 14 tasks] ●●●●●●●●●●●●──────…
#> Running `trauma_01_population()`  [Working on 6 of 14 tasks] ●●●●●●●●●●●●●●────…
#> Running `trauma_01_population()`  [Working on 7 of 14 tasks] ●●●●●●●●●●●●●●●●──…
#> Running `trauma_01_population()`  [Working on 8 of 14 tasks] ●●●●●●●●●●●●●●●●●●…
#> Running `trauma_01_population()`  [Working on 9 of 14 tasks] ●●●●●●●●●●●●●●●●●●…
#> Running `trauma_01_population()`  [Working on 10 of 14 tasks] ●●●●●●●●●●●●●●●●●…
#> Running `trauma_01_population()`  [Working on 11 of 14 tasks] ●●●●●●●●●●●●●●●●●…
#> Running `trauma_01_population()`  [Working on 12 of 14 tasks] ●●●●●●●●●●●●●●●●●…
#> Running `trauma_01_population()`  [Working on 13 of 14 tasks] ●●●●●●●●●●●●●●●●●…
#> Running `trauma_01_population()`  [Working on 14 of 14 tasks] ●●●●●●●●●●●●●●●●●…
#> 
#> 
#> 
#> ── Calculating Trauma-01 ──
#> 
#> 
#> ✔ Function completed in 0.21s.
#> 
#> # A tibble: 3 × 6
#>   measure   pop    numerator denominator  prop prop_label
#>   <chr>     <chr>      <int>       <int> <dbl> <chr>     
#> 1 Trauma-01 Adults         3           3     1 100%      
#> 2 Trauma-01 Peds           1           1     1 100%      
#> 3 Trauma-01 All            5           5     1 100%