Trauma-08 Calculation

This function calculates the Trauma-08 measure, which evaluates the completeness of pain scale documentation for patients experiencing traumatic injury. It determines the total population, adult population, and pediatric population meeting the criteria for the Trauma-08 measure.

trauma_08(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  disposition_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  esituation_02_col,
  eresponse_05_col,
  transport_disposition_col,
  evitals_06_col,
  evitals_14_col,
  evitals_23_col,
  ...
)

Arguments

df

A data frame or tibble containing EMS data with all relevant columns. Default is NULL.

patient_scene_table

A data frame or tibble containing only epatient and escene fields as a fact table. Default is NULL.

response_table

A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is NULL.

situation_table

A data frame or tibble containing only the esituation fields needed for this measure's calculations. Default is NULL.

disposition_table

A data frame or tibble containing only the edisposition fields needed for this measure's calculations. Default is NULL.

vitals_table

A data frame or tibble containing only the evitals fields needed for this measure's calculations. Default is NULL.

erecord_01_col

<'tidy-select'> A column specifying unique patient records.

incident_date_col

<'tidy-select'> A column indicating the incident date. Must be of class Date or POSIXct. Default is NULL.

patient_DOB_col

<'tidy-select'> A column indicating the patient's date of birth. Must be of class Date or POSIXct. Default is NULL.

epatient_15_col

<'tidy-select'> A column indicating the patient’s age in numeric form.

epatient_16_col

<'tidy-select'> A column specifying the unit of patient age (e.g., "Years", "Days").

esituation_02_col

<'tidy-select'> A column containing information about the nature of the patient’s condition (e.g., injury type).

eresponse_05_col

<'tidy-select'> A column specifying the type of response (e.g., 911 codes).

transport_disposition_col

<'tidy-select'> A column specifying transport disposition for the patient.

evitals_06_col

<'tidy-select'> A column containing systolic blood pressure (SBP) data from initial vital signs.

evitals_14_col

<'tidy-select'> A column containing respiratory rate data from initial vital signs.

evitals_23_col

<'tidy-select'> A column containing total Glasgow Coma Scale (GCS) scores from initial vital signs.

...

Additional arguments passed to the summarize_measure function.

Value

A tibble summarizing results for three population groups (All, Adults, and Peds) with the following columns:

pop: Population type (All, Adults, Peds). numerator: Count of incidents where the respiratory rate, SBP, and GCS vitals were taken. denominator: Total count of incidents. prop: Proportion of incidents where the respiratory rate, SBP, and GCS vitals were taken. prop_label: Proportion formatted as a percentage with a specified number of decimal places.

Data Assumptions

All vital signs are the initial vital signs.

This function accepts the input dataset contains the initial vital signs for respiratory rate, systolic blood pressure (SBP), and total Glasgow Coma Scale (GCS) score, respectively. If those values are not available, the user may pass all those values from the respective columns and the function will find the initial values.

Author

Nicolas Foss, Ed.D., MS