Trauma-14 Calculation

This function processes EMS data to generate a set of binary variables indicating whether specific trauma triage criteria are met. The output #' is a data frame enriched with these indicators for further analysis. The final outcome is whether or not the EMS record documents the use of #' a pre-hospital trauma activation.

Usage

trauma_14(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  vitals_table = NULL,
  exam_table = NULL,
  procedures_table = NULL,
  injury_table = NULL,
  disposition_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_dob_col = NULL,
  epatient_15_col,
  epatient_16_col,
  esituation_02_col,
  eresponse_05_col,
  eresponse_10_col,
  transport_disposition_col,
  edisposition_24_col,
  evitals_06_col,
  evitals_10_col,
  evitals_12_col,
  evitals_14_col,
  evitals_15_col,
  evitals_21_col,
  eexam_16_col,
  eexam_20_col,
  eexam_23_col,
  eexam_25_col,
  eprocedures_03_col,
  einjury_01_col,
  einjury_03_col,
  einjury_04_col,
  einjury_09_col,
  ...
)

Arguments

df

A data frame or tibble containing EMS data with all relevant columns.

patient_scene_table

A data frame or tibble containing fields from epatient and escene needed for this measure's calculations.

response_table

A data frame or tibble containing fields from eresponse needed for this measure's calculations.

situation_table

A data frame or tibble containing fields from esituation needed for this measure's calculations.

vitals_table

A data frame or tibble containing fields from evitals needed for this measure's calculations.

exam_table

A data frame or tibble containing fields from eexam needed for this measure's calculations.

procedures_table

A data frame or tibble containing fields from eprocedures needed for this measure's calculations.

injury_table

A data frame or tibble containing fields from einjury needed for this measure's calculations.

disposition_table

A data frame or tibble containing fields from edisposition needed for this measure's calculations.

erecord_01_col

The column representing the EMS record unique identifier.

incident_date_col

Column that contains the incident date. This defaults to NULL as it is optional in case not available due to PII restrictions.

patient_dob_col

Column that contains the patient's date of birth. This defaults to NULL as it is optional in case not available due to PII restrictions.

epatient_15_col

The column for patient age numeric value.

epatient_16_col

The column for patient age unit (e.g., "Years", "Months").

esituation_02_col

The column containing information on the presence of injury.

eresponse_05_col

The column representing the 911 response type.

eresponse_10_col

Column name containing scene delay information.

transport_disposition_col

The column for patient transport disposition.

edisposition_24_col

Column name containing pre-hospital trauma alert information.

evitals_06_col

Column name containing systolic blood pressure (SBP) values.

evitals_10_col

Column name containing heart rate values.

evitals_12_col

Column name containing pulse oximetry values.

evitals_14_col

Column name containing capillary refill information.

evitals_15_col

Column name containing respiratory effort values.

evitals_21_col

Column name containing Glasgow Coma Scale (GCS) Motor values.

eexam_16_col

Column name containing extremities assessment details.

eexam_20_col

Column name containing neurological assessment details.

eexam_23_col

Column name containing lung assessment details.

eexam_25_col

Column name containing chest assessment details.

eprocedures_03_col

Column name containing airway management or tourniquet usage details.

einjury_01_col

Column name containing injury cause details.

einjury_03_col

Column name containing trauma triage steps 1 and 2 information.

einjury_04_col

Column name containing trauma triage steps 3 and 4 information.

einjury_09_col

Column name containing fall height information.

...

Additional arguments passed to helper functions for further customization.

Value

A tibble summarizing results for three age groups (< 10 yrs, 10–65 yrs, and >= 65 yrs) with the following columns:

measure: The name of the measure being calculated. pop: Population type (< 10 yrs, 10–65 yrs, >= 65 yrs). numerator: Count of incidents where a pre-hospital trauma alert was called. denominator: Total count of incidents. prop: Proportion of incidents where a pre-hospital trauma alert was called. prop_label: Proportion formatted as a percentage with a specified number of decimal places.

Author

Nicolas Foss, Ed.D., MS