Trauma-14

This function processes EMS data to generate the population needed to calculated the Trauma-14 NEMSQA measure.

trauma_14_population(
  df = NULL,
  patient_scene_table = NULL,
  situation_table = NULL,
  response_table = NULL,
  disposition_table = NULL,
  vitals_table = NULL,
  exam_table = NULL,
  procedures_table = NULL,
  injury_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  esituation_02_col,
  eresponse_05_col,
  eresponse_10_col,
  transport_disposition_col,
  edisposition_24_col,
  evitals_06_col,
  evitals_10_col,
  evitals_12_col,
  evitals_14_col,
  evitals_15_col,
  evitals_21_col,
  eexam_16_col,
  eexam_20_col,
  eexam_23_col,
  eexam_25_col,
  eprocedures_03_col,
  einjury_01_col,
  einjury_03_col,
  einjury_04_col,
  einjury_09_col,
  ...
)

Arguments

df

A data frame or tibble containing EMS data with all relevant columns.

patient_scene_table

A data frame or tibble containing fields from epatient and escene needed for this measure's calculations.

situation_table

A data frame or tibble containing fields from esituation needed for this measure's calculations.

response_table

A data frame or tibble containing fields from eresponse needed for this measure's calculations.

disposition_table

A data frame or tibble containing fields from edisposition needed for this measure's calculations.

vitals_table

A data frame or tibble containing fields from evitals needed for this measure's calculations.

exam_table

A data frame or tibble containing fields from eexam needed for this measure's calculations.

procedures_table

A data frame or tibble containing fields from eprocedures needed for this measure's calculations.

injury_table

A data frame or tibble containing fields from einjury needed for this measure's calculations.

erecord_01_col

<'tidy-select'> The column representing the EMS record unique identifier.

incident_date_col

<'tidy-select'> The column indicating the incident date. Must be of class Date or similar.

patient_DOB_col

<'tidy-select'> The column representing the patient's date of birth. Must be of class Date or similar.

epatient_15_col

<'tidy-select'> The column for patient age numeric value.

epatient_16_col

<'tidy-select'> The column for patient age unit (e.g., "Years", "Months").

esituation_02_col

<'tidy-select'> The column containing information on the presence of injury.

eresponse_05_col

<'tidy-select'> The column representing the 911 response type.

eresponse_10_col

<'tidy-select'> Column name containing scene delay information.

transport_disposition_col

<'tidy-select'> The column for patient transport disposition.

edisposition_24_col

<'tidy-select'> Column name containing pre-hospital trauma alert information.

evitals_06_col

<'tidy-select'> Column name containing systolic blood pressure (SBP) values.

evitals_10_col

<'tidy-select'> Column name containing heart rate values.

evitals_12_col

<'tidy-select'> Column name containing pulse oximetry values.

evitals_14_col

<'tidy-select'> Column name containing capillary refill information.

evitals_15_col

<'tidy-select'> Column name containing respiratory effort values.

evitals_21_col

<'tidy-select'> Column name containing Glasgow Coma Scale (GCS) Motor values.

eexam_16_col

<'tidy-select'> Column name containing extremities assessment details.

eexam_20_col

<'tidy-select'> Column name containing neurological assessment details.

eexam_23_col

<'tidy-select'> Column name containing lung assessment details.

eexam_25_col

<'tidy-select'> Column name containing chest assessment details.

eprocedures_03_col

<'tidy-select'> Column name containing airway management or tourniquet usage details.

einjury_01_col

<'tidy-select'> Column name containing injury cause details.

einjury_03_col

<'tidy-select'> Column name containing trauma triage steps 1 and 2 information.

einjury_04_col

<'tidy-select'> Column name containing trauma triage steps 3 and 4 information.

einjury_09_col

<'tidy-select'> Column name containing fall height information.

...

Additional arguments passed to helper functions for further customization.

Value

#' A list that contains the following:

a tibble with counts for each filtering step,

a tibble for each population of interest

a tibble for the initial population

Data Assumptions

All vitals columns contain the full list of values entered for each record and columns.

Each eexam field is a list column containing all values entered per record that are also comma (or otherwise) separated.

The trauma triage fields are list columns containing all values entered per record that are also comma (or otherwise) separated.

einjury_01_col is a list column containing all values entered per record that are also comma (or otherwise) separated.

eresponse_10_col is a list column containing all values entered per record that are also comma (or otherwise) separated.

eprocedures_03_col is a list column containing all values entered per record that are also comma (or otherwise) separated.

Features

Handles missing or invalid date formats with error messaging.

Incorporates quasiquotation for flexible column referencing.

Creates reusable dimension tables for efficient filtering and summarization.

Author

Nicolas Foss, Ed.D., MS