TTR-01 Calculation

This function calculates the TTR-01 measure, which evaluates the completeness of vitals documentation for patients not experiencing cardiac arrest who were also not transported during a 911 response. It determines the total population, adult population, and pediatric population meeting the criteria for the TTR_01 measure.

Usage

ttr_01(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  disposition_table = NULL,
  vitals_table = NULL,
  arrest_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_dob_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  transport_disposition_col,
  earrest_01_col,
  evitals_06_col,
  evitals_07_col,
  evitals_10_col,
  evitals_12_col,
  evitals_14_col,
  evitals_23_col,
  evitals_26_col,
  ...
)

Arguments

df

A data frame or tibble containing the dataset to analyze. Default is NULL.

patient_scene_table

A data frame or tibble containing only epatient and escene fields as a fact table. Default is NULL.

response_table

A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is NULL.

disposition_table

A data frame or tibble containing only the edisposition fields needed for this measure's calculations. Default is NULL.

vitals_table

A data frame or tibble containing only the evitals fields needed for this measure's calculations. Default is NULL.

arrest_table

A data frame or tibble containing only the earrest fields needed for this measure's calculations. Default is NULL.

erecord_01_col

A column specifying unique patient records.

incident_date_col

Column that contains the incident date. This defaults to NULL as it is optional in case not available due to PII restrictions.

patient_dob_col

Column that contains the patient's date of birth. This defaults to NULL as it is optional in case not available due to PII restrictions.

epatient_15_col

A column indicating the patient’s age in numeric form.

epatient_16_col

A column specifying the unit of patient age (e.g., "Years", "Days").

eresponse_05_col

A column specifying the type of response (e.g., 911 codes).

transport_disposition_col

A column specifying transport disposition for the patient.

earrest_01_col

A column containing cardiac arrest data.

evitals_06_col

A column containing systolic blood pressure (SBP) data from initial vital signs.

evitals_07_col

A column containing diastolic blood pressure (DBP) data from initial vital signs.

evitals_10_col

A column containing heart rate data from initial vital signs.

evitals_12_col

A column containing spO2 data from the initial vital signs.

evitals_14_col

A column containing respiratory rate data from initial vital signs.

evitals_23_col

A column containing total Glasgow Coma Scale (GCS) scores from initial vital signs.

evitals_26_col

A column containing alert, verbal, painful, unresponsive (AVPU) vital signs.

...

Additional arguments passed to the summarize_measure function.

Value

A tibble summarizing results for three population groups (Adults, and Peds) with the following columns:

measure: The name of the measure being calculated. pop: Population type (Adults, Peds). numerator: Count of incidents where all applicable vital signs are taken. denominator: Total count of incidents. prop: Proportion of incidents where all applicable vital signs are taken. prop_label: Proportion formatted as a percentage with a specified number of decimal places.

Author

Nicolas Foss, Ed.D., MS