Respiratory-01 Calculation

The respiratory_01 function filters and analyzes data related to emergency 911 respiratory distress incidents, providing summary statistics for adult and pediatric populations. This function uses specific data columns for 911 response codes, primary and secondary impressions, and vital signs to calculate the proportion of cases with complete vital signs recorded, stratified by age.

respiratory_01(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  evitals_12_col,
  evitals_14_col,
  ...
)

Arguments

df

A data frame containing incident data with each row representing an observation.

patient_scene_table

A data.frame or tibble containing at least epatient and escene fields as a fact table.

response_table

A data.frame or tibble containing at least the eresponse fields needed for this measure's calculations.

situation_table

A data.frame or tibble containing at least the esituation fields needed for this measure's calculations.

vitals_table

A data.frame or tibble containing at least the evitals fields needed for this measure's calculations.

erecord_01_col

<'tidy-select'> Unique Patient ID

incident_date_col

<'tidy-select'> Date or POSIXct Column name for the Incident Date field.

patient_DOB_col

<'tidy-select'> Date or POSIXct Column name for epatient.17.

epatient_15_col

<'tidy-select'> Column giving the calculated age value.

epatient_16_col

<'tidy-select'> Column giving the provided age unit value.

eresponse_05_col

<'tidy-select'> Column name for 911 response codes (e.g., 2205001, 2205003, 2205009).

esituation_11_col

<'tidy-select'> Column name for primary impression codes related to respiratory distress.

esituation_12_col

<'tidy-select'> Column name for secondary impression codes related to respiratory distress.

evitals_12_col

<'tidy-select'> Column name for the first vital sign measurement.

evitals_14_col

<'tidy-select'> Column name for the second vital sign measurement.

...

arguments passed to dplyr::summarize().

Value

Returns a data frame summarizing the proportion of cases with complete vital sign data, divided by population

Assumptions

Assume data are already loaded Need to be a table where each row is 1 observation and each column is a feature or distinct datasets that can be referenced as unique columns this function will calculate an age in years this function also assumes that rows that are missing any value are NA, not the not known / not recorded values common to ImageTrend or the value codes that correspond to "not values". the function assumes that the primary/secondary impression fields have the ICD-10 code in them. The text description can be present, too, for reference. the function assumes that the eresponse.05 column has the codes in it, text can be present, too, for reference the function assumes that vitals in the vital signs columns are likely the first vital signs, or are a list column. This will give an indication of whether or not any vitals were taken. the esituation_12 is best as a list column of the secondary impressions entered the first argument is a dataframe, no joining is done. any joins to get vitals etc. will need to be done outside the function Grouping by specific attributes (e.g., region) can be performed inside this function by utilizing the .by argument passed via tidydots (i.e. ...) to dplyr::summarize.

Author

Nicolas Foss, Ed.D., MS