Respiratory-01 Populations

The respiratory_01_population function filters and analyzes data related to emergency 911 respiratory distress incidents, providing the adult, pediatric, and initial populations. This function uses specific data columns for 911 response codes, primary and secondary impressions, and vital signs to filter a dataset down to the populations of interest.

Usage

respiratory_01_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_dob_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  evitals_12_col,
  evitals_14_col
)

Arguments

df

A data frame containing incident data with each row representing an observation.

patient_scene_table

A data.frame or tibble containing at least epatient and escene fields as a fact table.

response_table

A data.frame or tibble containing at least the eresponse fields needed for this measure's calculations.

situation_table

A data.frame or tibble containing at least the esituation fields needed for this measure's calculations.

vitals_table

A data.frame or tibble containing at least the evitals fields needed for this measure's calculations.

erecord_01_col

<'tidy-select'> Unique Patient ID

incident_date_col

Column that contains the incident date. This defaults to NULL as it is optional in case not available due to PII restrictions.

patient_dob_col

Column that contains the patient's date of birth. This defaults to NULL as it is optional in case not available due to PII restrictions.

epatient_15_col

<'tidy-select'> Column giving the calculated age value.

epatient_16_col

<'tidy-select'> Column giving the provided age unit value.

eresponse_05_col

<'tidy-select'> Column name for 911 response codes (e.g., 2205001, 2205003, 2205009).

esituation_11_col

<'tidy-select'> Column name for primary impression codes related to respiratory distress.

esituation_12_col

<'tidy-select'> Column name for secondary impression codes related to respiratory distress.

evitals_12_col

<'tidy-select'> Column name for the first vital sign measurement.

evitals_14_col

<'tidy-select'> Column name for the second vital sign measurement.

Value

Returns a data frame summarizing the proportion of cases with complete vital sign data, divided by population

Author

Nicolas Foss, Ed.D., MS