Safety-02 Calculation

The safety_02 function calculates the Safety-02 metric, evaluating the proportion of emergency medical calls involving transport where no lights and sirens were used. This function categorizes the population into adult and pediatric groups based on their age, and summarizes results with a total population count as well.

safety_02(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  disposition_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  edisposition_18_col,
  edisposition_28_col,
  transport_disposition_cols,
  ...
)

Arguments

df

A data frame where each row is an observation, and each column represents a feature.

patient_scene_table

A data.frame or tibble containing only epatient and escene fields as a fact table.

response_table

A data.frame or tibble containing only the eresponse fields needed for this measure's calculations.

disposition_table

A data.frame or tibble containing only the edisposition fields needed for this measure's calculations.

incident_date_col

<'tidy-select'> Date or POSIXct Unquoted column name representing the date of the incident.

patient_DOB_col

<'tidy-select'> Date or POSIXct Unquoted column name for the patient's date of birth.

epatient_15_col

<'tidy-select'> Column giving the calculated age value.

epatient_16_col

<'tidy-select'> Column giving the provided age unit value.

eresponse_05_col

<'tidy-select'> Column giving response codes, identifying 911 responses.

edisposition_18_col

<'tidy-select'> Column giving transport mode descriptors, including possible lights-and-sirens indicators.

edisposition_28_col

<'tidy-select'> Column giving patient evaluation and care categories for the EMS response.

transport_disposition_cols

<'tidy-select'> One or more unquoted column names (such as edisposition.12, edisposition.30) containing transport disposition details.

...

Additional arguments for summary calculation, if needed.

Value

A tibble summarizing results for three age groups (< 10 yrs, 10–65 yrs, and >= 65 yrs) with the following columns:

pop: Population type (< 18 yrs, >= 18 yrs, all). numerator: Count of incidents from a 911 request during which lights and sirens were not used during patient transport. denominator: Total count of incidents. prop: Proportion of incidents from a 911 request during which lights and sirens were not used during patient transport. prop_label: Proportion formatted as a percentage with a specified number of decimal places.

Details

Assume data are already loaded Need to be a table where each row is 1 observation and each column is a feature or distinct datasets that can be referenced as unique columns this function will calculate an age in years this function also assumes that rows that are missing any value are NA, not the not known / not recorded values common to ImageTrend or the value codes that correspond to "not values". the function assumes that the eresponse.05 column has the codes in it, text can be present, too, for reference the function assumes that edisposition.18 is a list column or a column that has all text descriptors for additional transport mode descriptors. These can be separated by commas or other characters as long as all eresponse.18 values are present in one cell for each unique erecord.01 value. Codes can be present but will be ignored by the function. for the argument transport_disposition_cols, this argument can receive the unquoted column names of edisposition.12 and edisposition.30. One or both can be entered and the function will evaluate them. These columns are used to create a transport variable that is used to filter the table down furhter. AS such, these columns edisposition.12 and edisposition.30 must be list columns that and/or contain all values from each unique incident entered for each field. These can be comma separated values all in one cell to make the table tidy. the first argument is a dataframe, no joining is done. any joins to get vitals etc. will need to be done outside the function Grouping by specific attributes (e.g., region) can be performed inside this function by utilizing the .by argument passed via tidydots (i.e. ...) to dplyr::summarize.

Author

Nicolas Foss, Ed.D., MS