Seizure-02 Calculation

Calculates the NEMSQA Seizure-02 Measure.

Calculates age-based seizure metrics for a dataset. This function filters data for patients based on incident information, diagnoses, and administered medications to assess adherence to Seizure-02 metrics.

Usage

seizure_02(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  medications_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  emedications_03_col,
  ...
)

Arguments

df

A data frame where each row is an observation, containing all necessary columns for analysis.

patient_scene_table

A data frame or tibble containing only epatient and escene fields as a fact table. Default is NULL.

response_table

A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is NULL.

situation_table

A data.frame or tibble containing only the esituation fields needed for this measure's calculations. Default is NULL.

medications_table

A data.frame or tibble containing only the emedications fields needed for this measure's calculations. Default is NULL.

erecord_01_col

The column containing unique record identifiers for each encounter.

incident_date_col

Column that contains the incident date. This defaults to NULL as it is optional in case not available due to PII restrictions.

patient_DOB_col

Column that contains the patient's date of birth. This defaults to NULL as it is optional in case not available due to PII restrictions.

epatient_15_col

Column name for patient age in numeric form.

epatient_16_col

Column name for age unit (e.g., "Years" or "Months").

eresponse_05_col

Column name for response codes; "911" call codes are filtered.

esituation_11_col

Column name for primary impressions.

esituation_12_col

Column name for secondary impressions.

emedications_03_col

Column name for medications administered; ideally a list column or string with comma-separated values.

...

Additional arguments passed to dplyr::summarize.

Value

A tibble summarizing results for three population groups (All, Adults, and Peds) with the following columns:

measure: The name of the measure being calculated. pop: Population type (All, Adults, or Peds). numerator: Count of incidents where beta-agonist medications were administered. denominator: Total count of incidents. prop: Proportion of incidents involving beta-agonist medications. prop_label: Proportion formatted as a percentage with a specified number of decimal places.

Author

Nicolas Foss, Ed.D., MS