Syncope-01 Calculation

The syncope_01 function processes EMS dataset to identify potential syncope (fainting) cases based on specific criteria and calculates related ECG measures. This function dplyr::filters data for 911 response calls, assesses primary and associated symptoms for syncope, determines age-based populations (adult and pediatric), and aggregates results by unique patient encounters.

Usage

syncope_01(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_09_col,
  esituation_10_col,
  esituation_11_col,
  esituation_12_col,
  evitals_04_col,
  ...
)

Arguments

df

Main data frame containing EMS records.

patient_scene_table

A data frame or tibble containing only epatient and escene fields as a fact table. Default is NULL.

response_table

A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is NULL.

situation_table

A data.frame or tibble containing only the esituation fields needed for this measure's calculations. Default is NULL.

vitals_table

A data.frame or tibble containing only the evitals fields needed for this measure's calculations. Default is NULL.

erecord_01_col

The column containing unique record identifiers for each encounter.

incident_date_col

Column that contains the incident date. This defaults to NULL as it is optional in case not available due to PII restrictions.

patient_DOB_col

Column that contains the patient's date of birth. This defaults to NULL as it is optional in case not available due to PII restrictions.

epatient_15_col

Column representing the patient age (numeric).

epatient_16_col

Column for the patient age units (e.g., "Years", "Months").

eresponse_05_col

Column containing response type codes, specifically 911 codes.

esituation_09_col

Column with primary symptoms associated with the patient encounter.

esituation_10_col

Column with other associated symptoms.

esituation_11_col

Column for primary impression code.

esituation_12_col

Column for secondary impression codes.

evitals_04_col

Column with ECG information if available.

...

Additional arguments passed to dplyr::summarize for grouped summaries.

Value

A tibble summarizing results for three population groups (Adults, and Peds) with the following columns:

measure: The name of the measure being calculated. pop: Population type (Adults, Peds). numerator: Count of incidents where beta-agonist medications were administered. denominator: Total count of incidents. prop: Proportion of incidents involving beta-agonist medications. prop_label: Proportion formatted as a percentage with a specified number of decimal places.

Author

Nicolas Foss, Ed.D., MS