TBI-01 Calculation

This function screens for potential traumatic brain injury (TBI) cases based on specific criteria in a patient dataset. It produces a subset of the data with calculated variables for TBI identification.

Usage

tbi_01(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  disposition_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_DOB_col = NULL,
  epatient_15_col,
  epatient_16_col,
  eresponse_05_col,
  esituation_11_col,
  esituation_12_col,
  transport_disposition_col,
  evitals_06_col,
  evitals_12_col,
  evitals_16_col,
  evitals_23_col,
  evitals_26_col,
  ...
)

Arguments

df

A data frame or tibble containing the patient data.

patient_scene_table

A data frame or tibble containing only epatient and escene fields as a fact table. Default is NULL.

response_table

A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is NULL.

situation_table

A data.frame or tibble containing only the esituation fields needed for this measure's calculations. Default is NULL.

disposition_table

A data.frame or tibble containing only the edisposition fields needed for this measure's calculations. Default is NULL.

vitals_table

A data.frame or tibble containing only the evitals fields needed for this measure's calculations. Default is NULL.

erecord_01_col

Column name in df with the patient’s unique record ID.

incident_date_col

Column that contains the incident date. This defaults to NULL as it is optional in case not available due to PII restrictions.

patient_DOB_col

Column that contains the patient's date of birth. This defaults to NULL as it is optional in case not available due to PII restrictions.

epatient_15_col

Column name in df with the patient’s age value.

epatient_16_col

Column name in df with the patient’s age unit (e.g., years, months).

eresponse_05_col

Column name in df with response codes for the type of EMS call.

esituation_11_col

Column name in df with the primary provider impression.

esituation_12_col

Column name in df with the secondary provider impression.

transport_disposition_col

Column name in df with the transport disposition.

evitals_06_col

Column name in df with systolic blood pressure (SBP).

evitals_12_col

Column name in df with pulse oximetry values.

evitals_16_col

Column name in df with ETCO2 values. values.

evitals_23_col

Column name in df with Glasgow Coma Scale (GCS) scores.

evitals_26_col

Column name in df with AVPU (alert, verbal, painful, unresponsive) values.

...

Additional parameters passed to dplyr::summarize or other dplyr functions.

Value

A tibble summarizing results for three population groups (Adults, and Peds) with the following columns:

measure: The name of the measure being calculated. pop: Population type (Adults, Peds). numerator: Count of incidents where SP02, ETCO2, and SBP were all measured. denominator: Total count of incidents. prop: Proportion of incidents where SP02, ETCO2, and SBP were all measured. prop_label: Proportion formatted as a percentage with a specified number of decimal places.

Author

Nicolas Foss, Ed.D., MS