Trauma-08 Populations

Filters data down to the target populations for Trauma-08, and categorizes records to identify needed information for the calculations.

Identifies key categories to records that are 911 requests for patients with trauma during which GCS, systolic blood pressure, and respiratory rate are documented based on specific criteria and calculates related ECG measures. This function segments the data by age into adult and pediatric populations.

Usage

trauma_08_population(
  df = NULL,
  patient_scene_table = NULL,
  response_table = NULL,
  situation_table = NULL,
  disposition_table = NULL,
  vitals_table = NULL,
  erecord_01_col,
  incident_date_col = NULL,
  patient_dob_col = NULL,
  epatient_15_col,
  epatient_16_col,
  esituation_02_col,
  eresponse_05_col,
  transport_disposition_col,
  evitals_06_col,
  evitals_14_col,
  evitals_23_col
)

Arguments

df

A data frame or tibble containing EMS data with all relevant columns. Default is NULL.

patient_scene_table

A data frame or tibble containing only epatient and escene fields as a fact table. Default is NULL.

response_table

A data frame or tibble containing only the eresponse fields needed for this measure's calculations. Default is NULL.

situation_table

A data frame or tibble containing only the esituation fields needed for this measure's calculations. Default is NULL.

disposition_table

A data frame or tibble containing only the edisposition fields needed for this measure's calculations. Default is NULL.

vitals_table

A data frame or tibble containing only the evitals fields needed for this measure's calculations. Default is NULL.

erecord_01_col

A column specifying unique patient records.

incident_date_col

Column that contains the incident date. This defaults to NULL as it is optional in case not available due to PII restrictions.

patient_dob_col

Column that contains the patient's date of birth. This defaults to NULL as it is optional in case not available due to PII restrictions.

epatient_15_col

A column indicating the patient’s age in numeric form.

epatient_16_col

A column specifying the unit of patient age (e.g., "Years", "Days").

esituation_02_col

A column containing information about the nature of the patient’s condition (e.g., injury type).

eresponse_05_col

A column specifying the type of response (e.g., 911 codes).

transport_disposition_col

A column specifying transport disposition for the patient.

evitals_06_col

A column containing systolic blood pressure (SBP) data from initial vital signs.

evitals_14_col

A column containing respiratory rate data from initial vital signs.

evitals_23_col

A column containing total Glasgow Coma Scale (GCS) scores from initial vital signs.

Value

#' A list that contains the following:

a tibble with counts for each filtering step,

a tibble for each population of interest

a tibble for the initial population

Author

Nicolas Foss, Ed.D., MS